# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tộc Phả API
  version: v2
  description: |
    Hợp đồng API của Tộc Phả (bản chốt v2: một gia phả chung, hai vai trò Admin và User).
    File này viết tay và là **nguồn sự thật** (DECISIONS #70): frontend sinh kiểu TypeScript từ đây
    (`npm run gen:api`), backend phải khớp (test hợp đồng từ Đợt 26). Muốn đổi API thì sửa file này trước.

    Khởi tạo từ `/v3/api-docs` của backend Đợt 8, đã bỏ `/api/family/**`. Từ Đợt 39 frontend gọi backend thật cho
    mọi endpoint (đã gỡ lớp giả lập).

    Mọi API nghiệp vụ chỉ nhận tài khoản đã được duyệt (403 `ACCOUNT_NOT_APPROVED`), **trừ** các GET xem công khai
    (thành viên, người thân, cây, sự kiện, lịch, tổng quan; DECISIONS #88): khách và tài khoản chưa duyệt cũng gọi được,
    được đánh dấu `security: [{}, {bearerAuth: []}]`. Lỗi luôn là `ProblemDetail`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: 'Cùng origin với frontend (dev: proxy /api sang :8080)'
security:
  - bearerAuth: []
...
paths:
...
  /api/members:
    get:
      tags:
        - members
      summary: Danh sách thành viên
      description: |
        Không cần đăng nhập (khách và tài khoản chưa duyệt cũng xem được, DECISIONS #88). Tìm theo tên **không cần gõ dấu**. Các bộ lọc kết hợp bằng AND.
        `ageMin`/`ageMax` tính theo năm (còn sống: đến năm hiện tại; đã mất: đến năm mất dương). Người chưa rõ
        năm sinh, hoặc đã mất mà chưa rõ năm mất, không tính được tuổi nên bị loại khi có một trong hai tham số. `generation` và `onTree` lấy từ cây gia phả (cây trống thì
        không ai có đời và mọi người đều `onTree=false`).
      operationId: listMembers
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Tìm theo họ tên, không phân biệt hoa thường và dấu
          required: false
          schema:
            type: string
            maxLength: 100
        - name: sort
          in: query
          description: |
            `name` tên A-Z (theo cả chuỗi họ tên); `age` lớn tuổi trước, người chưa rõ năm sinh xếp cuối;
            `created` mới thêm trước; `generation` đời nhỏ trước, người chưa lên cây xếp cuối.
          required: false
          schema:
            type: string
            enum:
              - name
              - age
              - created
              - generation
            default: name
        - name: ageMin
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 150
        - name: ageMax
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 150
        - name: generation
          in: query
          description: Đời (độ sâu trên cây, đời 1 là gốc)
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: deceased
          in: query
          description: '`true` chỉ người đã mất, `false` chỉ người còn sống'
          required: false
          schema:
            type: boolean
        - name: onTree
          in: query
          description: '`true` chỉ người đã có trên cây, `false` chỉ người chưa có'
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
            maximum: 100
            minimum: 1
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
        - members
      summary: Thêm thành viên
      description: |
        Chỉ Admin. Chỉ `fullName` là bắt buộc. Các trường về việc đã mất (`deathSolar`, `deathLunar`,
        `memorialOverride`, `burialPlace`) chỉ hợp lệ khi `isDeceased = true`, nếu không thì 400 `VALIDATION_ERROR`.
        Ngày mất gửi **một trong hai** (`deathSolar` hoặc `deathLunar`); có cả hai thì dùng `deathSolar`. Khi có năm, máy chủ
        tự tính ngày ở lịch còn lại và trả cả hai; `deathLunar` không có năm thì `deathSolar` là `null`.
      operationId: createMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/members/{id}:
    get:
      tags:
        - members
      summary: Chi tiết thành viên
      description: |
        Không cần đăng nhập (khách và tài khoản chưa duyệt cũng xem được, DECISIONS #88). `phone` và `email` chỉ có trong body khi người gọi là Admin
        hoặc chính chủ hồ sơ (`user.member_id`); với người khác hai trường này **không xuất hiện**.
      operationId: getMember
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - members
      summary: Sửa thành viên
      description: |
        Admin sửa mọi hồ sơ. User đã liên kết "Tôi là ai" chỉ sửa hồ sơ của chính mình (`user.member_id`), có hiệu lực ngay,
        và **không** được đổi nhóm "đã mất" (`isDeceased`, `deathSolar`, `deathLunar`, `memorialOverride`, `burialPlace`):
        khác giá trị hiện có thì 403 `DEATH_FIELDS_ADMIN_ONLY`. Người khác thì 403 `FORBIDDEN`.
        Ảnh đại diện không đổi qua API này mà qua `/api/files/confirm`. Quy tắc dữ liệu như `POST /api/members`.
      operationId: updateMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - members
      summary: Xóa thành viên
      description: |
        Chỉ Admin. Bị chặn nếu người đó đang có trên cây (409 `MEMBER_ON_TREE`): phải gỡ khỏi cây trước.
        Khi xóa: xóa các dòng người thân liên quan, gỡ liên kết tài khoản, xóa tệp đính kèm; audit log giữ bản sao đầy đủ.
      operationId: deleteMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/members/{id}/relatives:
    get:
      tags:
        - relatives
      summary: Danh sách người thân của một hồ sơ
      description: |
        Không cần đăng nhập (khách và tài khoản chưa duyệt cũng xem được, DECISIONS #88). Một chiều: chỉ có các dòng do chủ hồ sơ hoặc Admin khai trong hồ sơ này,
        hồ sơ bên kia không tự có dòng ngược lại. Sắp theo thời điểm thêm.
      operationId: listRelatives
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Relative'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - relatives
      summary: Thêm người thân vào hồ sơ
      description: |
        Chủ hồ sơ (User đã liên kết, `user.member_id = id`) hoặc Admin; có hiệu lực ngay. `label` bắt buộc, cắt khoảng trắng hai đầu,
        tối đa 50 ký tự. Người thân phải là thành viên đã có. Lỗi: 400 `VALIDATION_ERROR` (nhãn, hoặc tự thêm chính mình),
        404 `MEMBER_NOT_FOUND`, 409 `RELATIVE_EXISTS` (người này đã có trong danh sách, muốn đổi nhãn thì sửa dòng cũ),
        403 `FORBIDDEN` (không phải chủ hồ sơ hay Admin).
      operationId: addRelative
      parameters:
        - $ref: '#/components/parameters/MemberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelativeInput'
      responses:
        '201':
          description: Đã thêm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relative'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/members/{id}/relatives/{relativeId}:
    put:
      tags:
        - relatives
      summary: Sửa nhãn của một người thân
      description: Chỉ đổi được `label` (quy tắc như khi thêm). Quyền như `POST`. 404 `RELATIVE_NOT_FOUND` khi dòng không thuộc hồ sơ này.
      operationId: updateRelative
      parameters:
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/RelativeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelativeLabelInput'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Relative'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - relatives
      summary: Xóa một người thân khỏi hồ sơ
      description: Quyền như `POST`. Chỉ xóa dòng trong hồ sơ này, thành viên được nhắc tới vẫn giữ nguyên.
      operationId: deleteRelative
      parameters:
        - $ref: '#/components/parameters/MemberId'
        - $ref: '#/components/parameters/RelativeId'
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    MemberId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    RelativeId:
      name: relativeId
      in: path
      required: true
      description: '`id` của dòng người thân (không phải id của thành viên được nhắc tới)'
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    ValidationError:
      description: 'Dữ liệu không hợp lệ (`VALIDATION_ERROR`, `MALFORMED_REQUEST`, lỗi lịch âm...)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Chưa đăng nhập, phiên hết hạn hoặc sai thông tin đăng nhập (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`...)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Không có quyền (`FORBIDDEN`), tài khoản chưa được duyệt (`ACCOUNT_NOT_APPROVED`) hoặc bị khóa (`ACCOUNT_LOCKED`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Không tìm thấy (`ACCOUNT_NOT_FOUND`, `MEMBER_NOT_FOUND`, `EVENT_NOT_FOUND`, `TREE_NODE_NOT_FOUND`, `RELATIVE_NOT_FOUND`, `LINK_REQUEST_NOT_FOUND`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: 'Xung đột trạng thái (`SELF_ACTION_FORBIDDEN`, `LAST_ADMIN`, `INVALID_ACCOUNT_STATE`, `MEMBER_ON_TREE`, `RELATIVE_EXISTS`, `MEMBER_ALREADY_LINKED`, `ACCOUNT_ALREADY_LINKED`, `LINK_REQUEST_EXISTS`, `LINK_REQUEST_NOT_PENDING`, `NOT_LINKED`, các mã cây `MEMBER_ALREADY_ON_TREE`, `TREE_*`)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    ProblemDetail:
      type: object
      description: Lỗi theo RFC 9457; `detail` là thông báo tiếng Việt hiển thị được cho người dùng.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Mã lỗi máy đọc được, ví dụ `ACCOUNT_NOT_APPROVED`, `MEMBER_NOT_FOUND`.
        errors:
          type: array
          description: Lỗi theo từng trường (chỉ có khi `code = VALIDATION_ERROR` hoặc lỗi lịch âm).
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required:
        - field
        - message
    SolarDate:
      type: object
      description: Ngày dương đầy đủ.
      properties:
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 31
      required:
        - year
        - month
        - day
    MemberLunarDate:
      type: object
      description: Ngày âm; `year` bỏ trống nghĩa là chỉ biết ngày/tháng (không tính "giỗ lần thứ N").
      properties:
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 30
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
        leap:
          type: boolean
          description: Ngày thuộc tháng nhuận
        year:
          type:
            - integer
            - 'null'
          format: int32
      required:
        - day
        - month
        - leap
    MemberBirth:
      type: object
      description: Ngày sinh; được phép chỉ có năm. `calendar` là lịch dùng để tính sinh nhật hằng năm của người này.
      properties:
        year:
          type:
            - integer
            - 'null'
          format: int32
        month:
          type:
            - integer
            - 'null'
          format: int32
          minimum: 1
          maximum: 12
        day:
          type:
            - integer
            - 'null'
          format: int32
          minimum: 1
          maximum: 31
        calendar:
          type: string
          enum:
            - SOLAR
            - LUNAR
          default: SOLAR
        leap:
          type: boolean
          description: Chỉ có nghĩa khi `calendar = LUNAR`
      required:
        - calendar
        - leap
    MemorialOverride:
      type: object
      description: Ngày giỗ ghi đè (âm lịch, không nhuận), dùng thay cho ngày mất khi gia đình cúng ngày khác.
      properties:
        day:
          type: integer
          format: int32
          minimum: 1
          maximum: 30
        month:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
      required:
        - day
        - month
    MemberSummary:
      type: object
      description: Dòng trong danh sách thành viên. Họ tên ghi nguyên văn, kể cả "Cụ", "Ông", "Bà", "(Tức ...)".
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        gender:
          type:
            - string
            - 'null'
          enum:
            - M
            - F
            - null
          description: Được để trống
        avatarUrl:
          type:
            - string
            - 'null'
        labels:
          type: array
          description: Nhãn đặc biệt, ví dụ "Liệt sỹ"
          items:
            type: string
        birthYear:
          type:
            - integer
            - 'null'
          format: int32
        isDeceased:
          type: boolean
        deathYear:
          type:
            - integer
            - 'null'
          format: int32
          description: Năm mất theo dương lịch (nếu biết)
        generation:
          type:
            - integer
            - 'null'
          format: int32
          description: Đời = độ sâu trên cây; `null` khi chưa có trên cây
        onTree:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - gender
        - avatarUrl
        - labels
        - birthYear
        - isDeceased
        - deathYear
        - generation
        - onTree
        - createdAt
    MemberDetail:
      description: Hồ sơ đầy đủ của một thành viên.
      allOf:
        - $ref: '#/components/schemas/MemberSummary'
        - type: object
          properties:
            tabooName:
              type:
                - string
                - 'null'
              description: Tên húy
            phone:
              type:
                - string
                - 'null'
              description: Chỉ có khi người gọi là Admin hoặc chính chủ; người khác không nhận trường này.
            email:
              type:
                - string
                - 'null'
              description: Chỉ có khi người gọi là Admin hoặc chính chủ; người khác không nhận trường này.
            biography:
              type:
                - string
                - 'null'
            birth:
              oneOf:
                - $ref: '#/components/schemas/MemberBirth'
                - type: 'null'
              description: '`null` khi chưa biết gì về ngày sinh'
            deathSolar:
              oneOf:
                - $ref: '#/components/schemas/SolarDate'
                - type: 'null'
              description: Ngày mất theo dương; tự tính từ ngày mất âm khi có năm, `null` khi chưa có.
            deathLunar:
              oneOf:
                - $ref: '#/components/schemas/MemberLunarDate'
                - type: 'null'
              description: Ngày mất theo âm (nguồn nhập của dữ liệu ban đầu); `null` khi chưa có.
            memorialOverride:
              oneOf:
                - $ref: '#/components/schemas/MemorialOverride'
                - type: 'null'
            burialPlace:
              type:
                - string
                - 'null'
            updatedAt:
              type: string
              format: date-time
          required:
            - tabooName
            - biography
            - birth
            - deathSolar
            - deathLunar
            - memorialOverride
            - burialPlace
            - updatedAt
    MemberInput:
      type: object
      description: |
        Nội dung hồ sơ khi thêm hoặc sửa (PUT thay toàn bộ, trường bỏ trống nghĩa là xóa giá trị).
        Họ tên ghi nguyên văn. Không có `avatarUrl`: ảnh đại diện đổi qua `/api/files/confirm`.
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 200
        gender:
          type:
            - string
            - 'null'
          enum:
            - M
            - F
            - null
        tabooName:
          type:
            - string
            - 'null'
          maxLength: 200
        labels:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 50
        biography:
          type:
            - string
            - 'null'
          maxLength: 5000
        phone:
          type:
            - string
            - 'null'
          maxLength: 30
        email:
          type:
            - string
            - 'null'
          maxLength: 254
        birth:
          oneOf:
            - $ref: '#/components/schemas/MemberBirth'
            - type: 'null'
        isDeceased:
          type: boolean
        deathSolar:
          oneOf:
            - $ref: '#/components/schemas/SolarDate'
            - type: 'null'
        deathLunar:
          oneOf:
            - $ref: '#/components/schemas/MemberLunarDate'
            - type: 'null'
        memorialOverride:
          oneOf:
            - $ref: '#/components/schemas/MemorialOverride'
            - type: 'null'
        burialPlace:
          type:
            - string
            - 'null'
          maxLength: 300
      required:
        - fullName
        - isDeceased
...
    MemberPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemberSummary'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
...
    Relative:
      type: object
      description: 'Một dòng trong danh sách người thân của hồ sơ: trong hồ sơ này, `relative` là "`label`". Ví dụ "Cụ Nguyễn Văn Uyên — cha".'
      properties:
        id:
          type: integer
          format: int64
        relative:
          $ref: '#/components/schemas/MemberSummary'
        label:
          type: string
          maxLength: 50
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - relative
        - label
        - createdAt
    RelativeInput:
      type: object
      properties:
        relativeMemberId:
          type: integer
          format: int64
          minimum: 1
        label:
          type: string
          description: Bắt buộc, cắt khoảng trắng hai đầu, tối đa 50 ký tự.
          maxLength: 50
      required:
        - relativeMemberId
        - label
    RelativeLabelInput:
      type: object
      properties:
        label:
          type: string
          description: Bắt buộc, cắt khoảng trắng hai đầu, tối đa 50 ký tự.
          maxLength: 50
      required:
        - label
...